+++ emanote/test/Emanote/Model/Link/RelSpec.hs
@@ -1,6 +1,9 @@
 module Emanote.Model.Link.RelSpec where
 
 import Emanote.Model.Link.Rel
+import Emanote.Model.Note qualified as MN
+import Emanote.Route.ModelRoute (mkModelRouteFromFilePath)
+import Emanote.Route.R (R (..))
 import Hedgehog
 import Relude
 import Test.Hspec
@@ -26,6 +29,25 @@ spec = do
       dropDotDot "foo/.." === ""
       dropDotDot "foo/bar/.." === "foo"
       dropDotDot "foo/bar/../.." === ""
+  describe "resolveLinkBaseFromFilePath" $ do
+    it "top-level files have no base" . hedgehog $ do
+      MN.resolveLinkBaseFromFilePath "index.md" === Nothing
+      MN.resolveLinkBaseFromFilePath "foo.md" === Nothing
+    it "subfolder/index.md uses subfolder as base (issue #608)" . hedgehog $ do
+      MN.resolveLinkBaseFromFilePath "subfolder/index.md"
+        === Just (R ("subfolder" :| []))
+    it "subfolder/foo.md uses subfolder as base" . hedgehog $ do
+      MN.resolveLinkBaseFromFilePath "subfolder/foo.md"
+        === Just (R ("subfolder" :| []))
+    it "a/b/index.md uses a/b as base" . hedgehog $ do
+      MN.resolveLinkBaseFromFilePath "a/b/index.md"
+        === Just (R ("a" :| ["b"]))
+  describe "parseUnresolvedRelTarget (issue #608)" $ do
+    it "relative link from subfolder/index.md resolves under subfolder/" . hedgehog $ do
+      let base = MN.resolveLinkBaseFromFilePath "subfolder/index.md"
+          got = fst <$> parseUnresolvedRelTarget base [] "foo.md"
+          want = URTResource <$> mkModelRouteFromFilePath "subfolder/foo.md"
+      got === want
 
 -- https://github.com/hedgehogqa/haskell-hedgehog/issues/121
 once :: PropertyT IO () -> Property